Langchain chat box/requirements.txt
langchain
openai
huggingface_hub
Python-dotenv
streamlit